requirements.txt
beautifulsoup4==4.5.1
lxml==3.3.5
python-amazon-product-api==0.2.8
wheel==0.24.0
